factorize BytesResourceReader for maps and worlds in a single module
