+++ src/map/asset.rs
@@ -1,10 +1,8 @@
 //! This module contains all [Asset]s definition.
 
-use std::io::{Cursor, Error as IoError, ErrorKind, Read};
+use std::io::ErrorKind;
 #[cfg(feature = "user_properties")]
 use std::ops::Deref;
-use std::path::Path;
-use std::sync::Arc;
 
 #[cfg(feature = "user_properties")]
 use bevy::reflect::TypeRegistryArc;
@@ -12,7 +10,7 @@ use bevy::reflect::TypeRegistryArc;
 #[cfg(feature = "user_properties")]
 use crate::properties::load::DeserializedMapProperties;
 
-use crate::cache::TiledResourceCache;
+use crate::{cache::TiledResourceCache, reader::BytesResourceReader};
 
 use bevy::{
     asset::{io::Reader, AssetLoader, AssetPath, LoadContext},
@@ -37,36 +35,6 @@ pub struct TiledMap {
     // The offset into the tileset_images for each tile id within each tileset.
     #[cfg(not(feature = "atlas"))]
     pub tile_image_offsets: HashMap<(usize, tiled::TileId), u32>,
-}
-
-struct BytesResourceReader<'a, 'b> {
-    bytes: Arc<[u8]>,
-    context: &'a mut LoadContext<'b>,
-}
-impl<'a, 'b> BytesResourceReader<'a, 'b> {
-    fn new(bytes: &'a [u8], context: &'a mut LoadContext<'b>) -> Self {
-        Self {
-            bytes: Arc::from(bytes),
-            context,
-        }
-    }
-}
-
-impl<'a> tiled::ResourceReader for BytesResourceReader<'a, '_> {
-    type Resource = Box<dyn Read + 'a>;
-    type Error = IoError;
-
-    fn read_from(&mut self, path: &Path) -> std::result::Result<Self::Resource, Self::Error> {
-        if let Some(extension) = path.extension() {
-            if extension == "tsx" {
-                let future = self.context.read_asset_bytes(path.to_path_buf());
-                let data = futures_lite::future::block_on(future)
-                    .map_err(|err| IoError::new(ErrorKind::NotFound, err))?;
-                return Ok(Box::new(Cursor::new(data)));
-            }
-        }
-        Ok(Box::new(Cursor::new(self.bytes.clone())))
-    }
 }
 
 pub(crate) struct TiledMapLoader {
